* int encoding fixed

+++ encoding/transaction/encoder.go
@@ -58,7 +58,7 @@ func (encoder *Encoder) Encode(v interface{}) error {
 
 	switch v := v.(type) {
 	case int:
-		return encoder.EncodeNumber(v)
+		return encoder.EncodeNumber(int32(v))
 	case int8:
 		return encoder.EncodeNumber(v)
 	case int16:
@@ -69,7 +69,7 @@ func (encoder *Encoder) Encode(v interface{}) error {
 		return encoder.EncodeNumber(v)
 
 	case uint:
-		return encoder.EncodeNumber(v)
+		return encoder.EncodeNumber(uint32(v))
 	case uint8:
 		return encoder.EncodeNumber(v)
 	case uint16:
